refactor(commands): migrate pair and run to internal/commands
Issue #282 step 4. Pure relocation:

- pair: PairCmd + pairingResult/requestPairing/pairWithRetry with the
  pairRetry* test seams
- run: RunCmd; TUI attach now via console.AttachConsoleTUI/AttachCodexTUI

Their tests move to internal/commands (exported names); the in-memory daemon
integration smoke stays in cmd/riffpad for now and moves with the dispatch.
