+++ apps/daemon/cmd/riffpad/main.go
@@ -25,7 +25,6 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/mdp/qrterminal/v3"
 	"golang.org/x/term"
 
 	"github.com/riffpad/riffpad/apps/daemon/internal/cliutil"
@@ -620,265 +619,11 @@ func forceKillDaemon(dataDir string) error {
 
 func statusCmd(base string) error { return commands.StatusCmd(base) }
 
-func pairCmd(base string, args []string) error {
-	fs := flag.NewFlagSet("pair", flag.ExitOnError)
-	local := fs.Bool("local", false, "mint a local-only code for the embedded UI at 8787, even when connected to a relay")
-	_ = fs.Parse(args)
-	pairURL := base + "/api/pairings"
-	if *local {
-		pairURL += "?local=1"
-	}
-	data, err := pairWithRetry(func() (pairingResult, error) { return requestPairing(pairURL) })
-	if err != nil {
-		return fmt.Errorf("%s: %w", t.T("daemon_not_reachable", base), err)
-	}
-	if data.Code == "" {
-		if data.ErrorCode == "relay_auth_expired" {
-			return fmt.Errorf("%s", t.T("pair_login_expired"))
-		}
-		if data.Error != "" {
-			return fmt.Errorf("%s", data.Error)
-		}
-		return fmt.Errorf("%s", t.T("pair_failed_status", data.Status))
-	}
-	if data.Local {
-		if !*local {
-			return fmt.Errorf("%s", t.T("pair_requires_login"))
-		}
-		// Local mode: the URL points at 127.0.0.1 and is only meaningful in a
-		// browser on this machine, so a QR code (which implies scanning with
-		// another device) would be misleading — print the URL instead.
-		fmt.Println(t.T("pair_local", data.Code, data.URL))
-		return nil
-	}
-	fmt.Println(t.T("pair_code", data.Code))
-	qrterminal.GenerateWithConfig(data.URL, qrterminal.Config{
-		Level:     qrterminal.L,
-		Writer:    os.Stdout,
-		BlackChar: qrterminal.BLACK,
-		WhiteChar: qrterminal.WHITE,
-		QuietZone: 1,
-	})
-	return nil
-}
-
-type pairingResult struct {
-	Status    int    `json:"-"`
-	Code      string `json:"code"`
-	URL       string `json:"url"`
-	Local     bool   `json:"local"`
-	Error     string `json:"error"`
-	ErrorCode string `json:"errorCode"`
-}
-
-func requestPairing(pairURL string) (pairingResult, error) {
-	var data struct {
-		Code      string `json:"code"`
-		URL       string `json:"url"`
-		Local     bool   `json:"local"`
-		Error     string `json:"error"`
-		ErrorCode string `json:"errorCode"`
-	}
-	resp, err := daemonDo(nil, http.MethodPost, pairURL, nil)
-	if err != nil {
-		return pairingResult{}, err
-	}
-	defer resp.Body.Close()
-	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return pairingResult{}, err
-	}
-	return pairingResult{
-		Status:    resp.StatusCode,
-		Code:      data.Code,
-		URL:       data.URL,
-		Local:     data.Local,
-		Error:     data.Error,
-		ErrorCode: data.ErrorCode,
-	}, nil
-}
-
-// pairWithRetry treats "host offline" as transient and keeps retrying until
-// the daemon registers with the relay or the deadline passes. All other
-// responses (success, auth expiry, host not found, etc.) return immediately.
-func pairWithRetry(post func() (pairingResult, error)) (pairingResult, error) {
-	deadline := time.Now().Add(pairRetryMaxWait)
-	waited := false
-	for {
-		data, err := post()
-		if err != nil {
-			return data, err
-		}
-		if data.Code != "" || data.ErrorCode != "" || data.Error != "host offline" || !time.Now().Before(deadline) {
-			return data, nil
-		}
-		if !waited {
-			fmt.Fprintln(os.Stderr, t.T("pair_waiting_host"))
-			waited = true
-		}
-		time.Sleep(pairRetryDelay)
-	}
-}
+func pairCmd(base string, args []string) error { return commands.PairCmd(base, args) }
 
 func sessionsCmd(base string) error { return commands.SessionsCmd(base) }
 
-func runCmd(args []string, base string) error {
-	fs := flag.NewFlagSet("run", flag.ExitOnError)
-	name := fs.String("name", "", "session name")
-	prompt := fs.String("prompt", "", "initial prompt")
-	cwd := fs.String("cwd", "", "working directory")
-	cli := fs.String("cli", "", "agent CLI (claude|kimi|codex)")
-	_ = fs.Parse(args)
-	rest := fs.Args()
-	if len(rest) > 1 {
-		return fmt.Errorf("unexpected arguments: %v", rest)
-	}
-	if len(rest) == 1 {
-		*cli = rest[0]
-	}
-	if *cli == "" {
-		*cli = "claude"
-	}
-	if *cwd == "" {
-		// Default to the directory where the user ran the command, not the
-		// daemon's own cwd (the daemon may have been started elsewhere).
-		if wd, err := os.Getwd(); err == nil {
-			*cwd = wd
-		}
-	}
-	// Resolve the CLI binary from the user's interactive PATH so sessions
-	// work even when the daemon runs under systemd with a minimal PATH.
-	// Empty is fine: the daemon falls back to its own PATH lookup.
-	binary := ""
-	if p, err := exec.LookPath(*cli); err == nil {
-		binary = p
-	}
-	body, _ := json.Marshal(map[string]string{
-		"name": *name, "prompt": *prompt, "cwd": *cwd, "cli": *cli, "binary": binary,
-	})
-	resp, err := daemonDo(nil, http.MethodPost, base+"/api/sessions", bytes.NewReader(body))
-	if err != nil {
-		return fmt.Errorf("%s: %w", t.T("daemon_not_reachable", base), err)
-	}
-	defer resp.Body.Close()
-	var data struct {
-		ID    string `json:"id"`
-		URL   string `json:"url"`
-		CLI   string `json:"cli"`
-		Error string `json:"error"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&data); err != nil {
-		return err
-	}
-	if data.ID == "" {
-		if data.Error != "" {
-			return fmt.Errorf("%s", data.Error)
-		}
-		return fmt.Errorf("%s", t.T("run_failed_status", resp.StatusCode))
-	}
-	if data.CLI == "codex" {
-		return console.AttachCodexTUI(base, data.ID)
-	}
-	if data.CLI == "demo" {
-		fmt.Println(t.T("session_url", data.ID, data.URL))
-		return nil
-	}
-	if data.CLI == "claude" || data.CLI == "kimi" {
-		if runtime.GOOS == "windows" {
-			fmt.Println(t.T("session_url", data.ID, data.URL))
-			return nil
-		}
-		return console.AttachConsoleTUI(base, data.ID, data.CLI)
-	}
-	fmt.Println(t.T("session_url", data.ID, data.URL))
-	return nil
-}
-
-// attachCodexTUI waits for the daemon's Codex app-server thread and then runs
-// `codex resume --remote` in the foreground so the TUI stays in the user's
-// terminal (no-silent hosting). Ctrl-C exits the TUI; the daemon session
-// remains available from the phone.
-func attachCodexTUI(base, sessionID string) error {
-	fmt.Println(t.T("codex_tui_starting"))
-	client := &http.Client{Timeout: 60 * time.Second}
-	resp, err := daemonDo(client, http.MethodGet, base+"/api/sessions/"+sessionID+"/connect", nil)
-	if err != nil {
-		return fmt.Errorf("%s", t.T("codex_tui_wait_failed", err))
-	}
-	defer resp.Body.Close()
-	var info struct {
-		Socket   string `json:"socket"`
-		ThreadID string `json:"threadId"`
-	}
-	if err := json.NewDecoder(resp.Body).Decode(&info); err != nil {
-		return err
-	}
-	if resp.StatusCode != http.StatusOK || info.Socket == "" || info.ThreadID == "" {
-		return fmt.Errorf("%s", t.T("codex_tui_not_ready", resp.StatusCode))
-	}
-	codexBin, err := exec.LookPath("codex")
-	if err != nil {
-		return fmt.Errorf("%s", t.T("codex_not_found", err))
-	}
-	cmd := exec.Command(codexBin, "resume", "--remote", "unix://"+info.Socket, info.ThreadID)
-	cmd.Stdin = os.Stdin
-	cmd.Stdout = os.Stdout
-	cmd.Stderr = os.Stderr
-
-	// Ctrl+C (SIGINT) is delivered to the whole foreground process group,
-	// i.e. both codex and this CLI. Without handling it, Go kills the CLI
-	// immediately and the session cleanup below never runs — the daemon
-	// session would stay alive and stay remote-controllable. So capture the
-	// signal, let codex exit normally (or kill it after a timeout), then
-	// close the session.
-	sigCh := make(chan os.Signal, 1)
-	signal.Notify(sigCh, os.Interrupt, syscall.SIGTERM, syscall.SIGHUP)
-	exited := make(chan struct{})
-	go func() {
-		select {
-		case <-sigCh:
-			select {
-			case <-exited:
-			case <-time.After(5 * time.Second):
-				if cmd.Process != nil {
-					_ = cmd.Process.Kill()
-				}
-			}
-		case <-exited:
-		}
-	}()
-	// Lease heartbeat: renew every 5s so the daemon can close the session if
-	// this CLI dies without running cleanup (SIGKILL, panic, network loss).
-	hbStop := make(chan struct{})
-	defer close(hbStop)
-	go func() {
-		ticker := time.NewTicker(5 * time.Second)
-		defer ticker.Stop()
-		for {
-			select {
-			case <-ticker.C:
-				if resp, err := daemonDo(nil, http.MethodPost, base+"/api/sessions/"+sessionID+"/heartbeat", nil); err == nil {
-					_ = resp.Body.Close()
-				}
-			case <-hbStop:
-				return
-			}
-		}
-	}()
-	runErr := cmd.Run()
-	close(exited)
-	if runErr != nil {
-		fmt.Fprintln(os.Stderr, t.T("codex_tui_exit_error"), runErr)
-	}
-	// The user exited the TUI — per the no-silent-hosting convention, exiting
-	// means exiting. Close the daemon session so it disappears from the client
-	// and cannot be remote-controlled anymore. Users who want a persistent
-	// session should run riffpad inside tmux themselves.
-	if resp, err := daemonDo(nil, http.MethodPost, base+"/api/sessions/"+sessionID+"/stop", nil); err == nil {
-		_ = resp.Body.Close()
-	}
-	fmt.Printf("%s\n", t.T("codex_tui_exited", sessionID))
-	return nil
-}
+func runCmd(args []string, base string) error { return commands.RunCmd(args, base) }
 
 func logsCmd(dataDir string) error { return commands.LogsCmd(dataDir) }
